+++ clients/vscode-hlasmplugin/src/test/suite/hlasmDownloadCommands.test.ts
@@ -18,9 +18,10 @@ import * as fsp from 'fs/promises';
 import * as path from 'path';
 import * as os from 'os';
 
-import { downloadDependenciesWithClient, extractDsn, gatherDownloadList, JobDescription, replaceVariables, adjustJobHeader, unterse } from '../../hlasmDownloadCommands';
+import { downloadDependenciesWithClient, extractDsn, gatherDownloadList, JobDescription, replaceVariables, adjustJobHeader, unterse, zoweJobMapper, ftpJobMapper, extractJobId } from '../../hlasmDownloadCommands';
 import { isCancellationError } from '../../helpers';
 import { arrayFromHex } from '../../tools';
+import type { FileInfo } from 'basic-ftp';
 
 suite('HLASM Download data sets', () => {
     const getClient = (listResponses: JobDescription[][]) => {
@@ -28,14 +29,14 @@ suite('HLASM Download data sets', () => {
             setListMaskCalls: new Array<string>(),
             listCalls: 0,
             jcls: new Array<string>(),
-            downloadRequests: new Array<{ id: string, spoolFile: number }>(),
+            downloadRequests: new Array<{ id: string, spoolFiles: string }>(),
             disposeCalls: 0,
             nextJobId: 0,
 
             dispose() { ++this.disposeCalls },
-            async download(target: Writable, id: string, spoolFile: number) {
-                if (target instanceof Writable) {
-                    this.downloadRequests.push({ id, spoolFile });
+            async download(target: Writable, job: JobDescription) {
+                if (target instanceof Writable && job.spoolFiles) {
+                    this.downloadRequests.push({ id: job.id, spoolFiles: job.spoolFiles });
                 }
                 else
                     assert.fail("Writable stream expected");
@@ -87,7 +88,7 @@ suite('HLASM Download data sets', () => {
     test('Simple jobcard', async () => {
         const client = getClient([
             [],
-            [{ jobname: "JOBNAME", id: "JOBID0", details: "RC=0000 3 spool files" }]
+            [{ jobname: "JOBNAME", id: "JOBID0", rc: 0, spoolFiles: "3" }]
         ]);
         const io = getIoOps();
         const stages = getStageCounter();
@@ -103,7 +104,7 @@ suite('HLASM Download data sets', () => {
         );
 
         assert.strictEqual(client.disposeCalls, 1);
-        assert.deepStrictEqual(client.downloadRequests, [{ id: "JOBID0", spoolFile: 3 }]);
+        assert.deepStrictEqual(client.downloadRequests, [{ id: "JOBID0", spoolFiles: "3" }]);
         assert.strictEqual(client.jcls.length, 1);
         assert.ok(client.jcls[0].startsWith("//JOBNAME JOB 1"));
         assert.notEqual(client.jcls[0].indexOf("DSN=A.B"), -1);
@@ -116,7 +117,7 @@ suite('HLASM Download data sets', () => {
 
     test('Jobcard pattern', async () => {
         const client = getClient([
-            [{ jobname: "JOBNAME0", id: "JOBID0", details: "RC=0000 3 spool files" }]
+            [{ jobname: "JOBNAME0", id: "JOBID0", rc: 0, spoolFiles: "3" }]
         ]);
         const io = getIoOps();
         const stages = getStageCounter();
@@ -132,7 +133,7 @@ suite('HLASM Download data sets', () => {
         );
 
         assert.strictEqual(client.disposeCalls, 1);
-        assert.deepStrictEqual(client.downloadRequests, [{ id: "JOBID0", spoolFile: 3 }]);
+        assert.deepStrictEqual(client.downloadRequests, [{ id: "JOBID0", spoolFiles: "3" }]);
         assert.strictEqual(client.jcls.length, 1);
         assert.ok(client.jcls[0].startsWith("//JOBNAME0 JOB 1"));
         assert.notEqual(client.jcls[0].indexOf("DSN=A.B"), -1);
@@ -145,7 +146,7 @@ suite('HLASM Download data sets', () => {
 
     test('Cancelled', async () => {
         const client = getClient([
-            [{ jobname: "JOBNAME0", id: "JOBID0", details: "RC=0000 3 spool files" }]
+            [{ jobname: "JOBNAME0", id: "JOBID0", rc: 0, spoolFiles: "3" }]
         ]);
         const io = getIoOps();
         const stages = getStageCounter();
@@ -177,10 +178,10 @@ suite('HLASM Download data sets', () => {
 
     test('Multiple data sets', async () => {
         const client = getClient([
-            [{ jobname: "JOBNAME0", id: "JOBID0", details: "RC=0000 3 spool files" }],
+            [{ jobname: "JOBNAME0", id: "JOBID0", rc: 0, spoolFiles: "3" }],
             [
-                { jobname: "JOBNAME0", id: "JOBID0", details: "RC=0000 3 spool files" },
-                { jobname: "JOBNAME1", id: "JOBID1", details: "RC=0000 6 spool files" }
+                { jobname: "JOBNAME0", id: "JOBID0", rc: 0, spoolFiles: "3" },
+                { jobname: "JOBNAME1", id: "JOBID1", rc: 0, spoolFiles: "6" }
             ]
         ]);
         const io = getIoOps();
@@ -200,7 +201,7 @@ suite('HLASM Download data sets', () => {
         );
 
         assert.strictEqual(client.disposeCalls, 1);
-        assert.deepStrictEqual(client.downloadRequests, [{ id: "JOBID0", spoolFile: 3 }, { id: "JOBID1", spoolFile: 6 }]);
+        assert.deepStrictEqual(client.downloadRequests, [{ id: "JOBID0", spoolFiles: "3" }, { id: "JOBID1", spoolFiles: "6" }]);
         assert.strictEqual(client.jcls.length, 2);
         assert.ok(client.jcls[0].startsWith("//JOBNAME0 JOB 1"));
         assert.notEqual(client.jcls[0].indexOf("DSN=A.B"), -1);
@@ -215,7 +216,7 @@ suite('HLASM Download data sets', () => {
 
     test('Failed job', async () => {
         const client = getClient([
-            [{ jobname: "JOBNAME", id: "JOBID0", details: "RC=0008 3 spool files" }]
+            [{ jobname: "JOBNAME", id: "JOBID0", rc: 8, spoolFiles: "3" }]
         ]);
         const io = getIoOps();
         const stages = getStageCounter();
@@ -309,4 +310,54 @@ suite('HLASM Download data sets', () => {
 
         assert.match(data, / {80}\r?\n/);
     }).timeout(2000);
+
+    test('Zowe job mapper', () => {
+        const jobs = zoweJobMapper("DDNAME", [{
+            jobname: "JOB1",
+            jobid: "JOB123",
+            retcode: "U0001",
+        }, {
+            jobname: "JOB2",
+            jobid: "JOB321",
+            retcode: "CC 0004",
+        }, {
+            jobname: "JOB3",
+            jobid: "JOB789",
+            retcode: undefined,
+        }]);
+
+        assert.deepStrictEqual(jobs, [
+            { jobname: "JOB1", id: "JOB123", rc: undefined, spoolFiles: undefined, },
+            { jobname: "JOB2", id: "JOB321", rc: 4, spoolFiles: "DDNAME", },
+            { jobname: "JOB3", id: "JOB789", rc: undefined, spoolFiles: undefined, },
+        ]);
+    });
+
+    test('Ftp job mapper', () => {
+        assert.deepStrictEqual(
+            ftpJobMapper({ name: "JOB1 JOB123 ABEND=222 4 spool files", } as unknown as FileInfo),
+            { jobname: "JOB1", id: "JOB123", rc: undefined, spoolFiles: undefined, }
+        );
+        assert.deepStrictEqual(
+            ftpJobMapper({ name: "JOB2 JOB321 RC=0004 2 spool files", } as unknown as FileInfo),
+            { jobname: "JOB2", id: "JOB321", rc: 4, spoolFiles: "2", },
+        );
+    });
+
+    test('Ftp jobid extractor', () => {
+        assert.deepStrictEqual(extractJobId("It is known to JES as JOB12345"), "JOB12345");
+    });
+
+    test('Job tools invalid', () => {
+        try {
+            ftpJobMapper({ name: "Err", } as unknown as FileInfo);
+            assert.ok(false);
+        } catch (_) {
+        }
+        try {
+            extractJobId("Err");
+            assert.ok(false);
+        } catch (_) {
+        }
+    });
 });