+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -5,6 +5,7 @@
 #### Added
 - Support file change notification registration
 - Validate operand length in machine instructions
+- Provide bulk copybook download via Zowe Explorer
 
 #### Fixed
 - Logs produced by language server in WebAssembly mode are not visible